fix: resolve double call to onExecute
